Change hardcoded timer irq from 2 to 0
The irq src for the timer is 0 which will usually be mapped as gsi 2.
UVMM does not do this mapping, but hermit assumes this mapping to be
present, as the value is hardcoded. A proper fix would be to read in the
mappings from the MADT ACPI table. Alternatively UVMM could do this
mapping and write it into the MADT.